+++ app/actions.ts
@@ -330,31 +330,6 @@ export async function fetchUserPageTitle(player: string | number) {
   return null;
 }
 
-export async function fetchPlayerStats(player: string | number, params) {
-  const osuMode =
-    (await cookies().get('OTR-user-selected-osu-mode')?.value as Ruleset | undefined) ?? Ruleset.Osu;
-
-  const queryCheck = await UserpageQuerySchema.safeParse({
-    time: params?.time,
-  });
-
-  const wrapper = new PlayersWrapper(apiWrapperConfiguration);
-
-  let minDate = undefined;
-  if (queryCheck.success && queryCheck.data.time) {
-    minDate = new Date();
-    minDate.setDate(minDate.getDate() - params?.time);
-  }
-
-  const statsResponse = await wrapper.getStats({ key: player as string, ruleset: osuMode, dateMin: minDate });
-
-  if (statsResponse.status != 200) {
-    return statsResponse.status === 404 ? notFound() : redirect('/');
-  }
-
-  return statsResponse.result;
-}
-
 export async function paginationParamsToURL(params: {}) {
   let url = '';
